Implement clickListener on button and rename incorrect package name import
