+++ app/src/main/java/com/oakkub/survey/ui/surveys/list/adapter/content/viewholders/SurveysContentItemViewHolder.kt
@@ -1,11 +1,12 @@
-package com.oakkub.survey.ui.surveys.adapter.content.viewholders
+package com.oakkub.survey.ui.surveys.list.adapter.content.viewholders
 
 import android.support.v7.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 import com.eggdigital.trueyouedc.extensions.views.inflate
 import com.oakkub.survey.R
-import com.oakkub.survey.ui.surveys.adapter.content.SurveysItemAdapterModel
+import com.oakkub.survey.ui.surveys.list.adapter.content.OnSurveyItemClickListener
+import com.oakkub.survey.ui.surveys.list.adapter.content.SurveysItemAdapterModel
 import kotlinx.android.extensions.LayoutContainer
 import kotlinx.android.synthetic.main.item_surveys_content.*
 
@@ -14,7 +15,11 @@ import kotlinx.android.synthetic.main.item_surveys_content.*
  */
 class SurveysContentItemViewHolder private constructor(override val containerView: View) : RecyclerView.ViewHolder(containerView), LayoutContainer {
 
-    fun bind(item: SurveysItemAdapterModel.Item) {
+    fun bind(item: SurveysItemAdapterModel.Item, onSurveyItemClickListener: OnSurveyItemClickListener) {
+        surveysItemButton.setOnClickListener {
+            onSurveyItemClickListener(item)
+        }
+
         val response = item.surveyResponse
         surveysItemCoverImageView.setImageURI(response.highResCoverImageUrl)
         surveysItemTitleTextView.text = response.title
