+++ app/src/main/java/com/oakkub/survey/ui/surveys/list/adapter/content/viewholders/SurveysProgressItemViewHolder.kt
@@ -1,4 +1,4 @@
-package com.oakkub.survey.ui.surveys.adapter.content.viewholders
+package com.oakkub.survey.ui.surveys.list.adapter.content.viewholders
 
 import android.support.v7.widget.RecyclerView
 import android.view.View
